+++ capstone/src/main/java/com/google/univiz/api/CollegeDataApiImpl.java
@@ -1,0 +1,3 @@
+package com.google.univiz.api;
+
+final class CollegeDataApiImpl implements CollegeDataApi {}
