+++ capstone/src/main/java/com/google/univiz/api/SuggestionDataApi.java
@@ -1,4 +1,4 @@
-package com.google.univiz.api.resource;
+package com.google.univiz.api;
 
 import java.util.List;
