+++ capstone/src/main/java/com/google/univiz/api/representation/CollegeData.java
@@ -1,6 +1,5 @@
-package com.google.univiz.scorecard;
+package com.google.univiz.api.representation;
 
-import com.google.univiz.api.resource.CollegeId;
 import com.google.auto.value.AutoValue;
 
 /** CollegeData is a class representing a single university/college/institution. */
