+++ capstone/src/main/java/com/google/univiz/api/representation/CollegeId.java
@@ -1,4 +1,4 @@
-package com.google.univiz.api.resource;
+package com.google.univiz.api.representation;
 
 import com.google.auto.value.AutoValue;
 import com.ryanharter.auto.value.gson.GenerateTypeAdapter;
