+++ capstone/src/main/java/com/google/univiz/api/representation/MapsData.java
@@ -1,4 +1,4 @@
-package com.google.univiz.api.resource;
+package com.google.univiz.api.representation;
 
 /** The MapsData class will contain all college data that is specific to the Maps feature */
 public class MapsData {}
