+++ capstone/src/main/java/com/google/univiz/api/representation/TuitionStats.java
@@ -1,4 +1,4 @@
-package com.google.univiz.api.dataviz;
+package com.google.univiz.api.representation;
 
 import com.google.auto.value.AutoValue;
