+++ capstone/src/main/java/com/google/univiz/api/resource/MapResourceImpl.java
@@ -1,0 +1,3 @@
+package com.google.univiz.api.resource;
+
+public class MapResourceImpl implements MapResource {}
