+++ capstone/src/main/java/com/google/univiz/api/resource/SearchResourceImpl.java
@@ -1,0 +1,3 @@
+package com.google.univiz.api.resource;
+
+public class SearchResourceImpl implements SearchResource {}
