+++ capstone/src/test/java/com/google/univiz/scorecard/ScorecardConverterTest.java
@@ -3,7 +3,8 @@ package com.google.univiz.scorecard;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
 
-import com.google.univiz.api.resource.CollegeId;
+import com.google.univiz.api.representation.CollegeData;
+import com.google.univiz.api.representation.CollegeId;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
